svelte-dialog(refactor): remove createDialog
